OCR/requirements.txt
marker-pdf>=0.3.0
rich>=13.0.0

